// src/dtos/dashboard.dto.ts
export interface DashboardStatsDto {
  totalCompanies: number | null;
  totalBranches: number | null;
  totalUsers: number | null;
  totalDoctors: number | null;
  totalPatients: number | null;
  totalAppointments: number | null;
  totalPayments: number | null;
  totalHomeVisits: number | null;
}